refactor(codegen): extract number literal serialization (#24888)
Number literal serialization is coupled to `Codegen`, even though the formatter produces one string before codegen writes it. That coupling makes the formatting policy difficult to reuse without mixing a move with an implementation rewrite.

Extract the existing algorithm into a codegen-local callback helper. Generated output and allocation behavior are unchanged.

This is the first PR in the stack. #24876 moves the helper unchanged into `oxc_ecmascript`.

Verified with `cargo test -p oxc_codegen`, Clippy, formatting, and the combined stack with `just ready`.

AI assistance: Codex was used to implement and verify this change.
